Add files via upload
Please check campuswire for the image that might be helpful.
Here are my posts on the campuswire:

I have just updated the Grading.java and the OptionList.java for the students to upload any files of answer.

See submission only works right now when the student is checking the submission right after he or she submitted the file.

It needs the old version of CreateAccount.java to run and you might need to paste the users' information into AccountsData.txt and change the CreateAccount.java to the updated version to let the teacher see the file submission without cleanning the users' data.

Manually grading function is not added yet. I would try more on that after I came back from my next course. However, as soon as the CreateAccount.java can work as intended without cleanning users' data each time(this occurs in the old version) or unable to create an account(this opccurs in the latest version), the updated OptionList.java should work out fine.